before week7

+++ router/account.js
@@ -3,8 +3,8 @@ const inputCheck = require("../module/inputCheck.js");
 
 const {Client} = require("pg")
 const db = require('../database.js');
-const verify = require("../module/verify.js")
-
+const verify = require("../module/verify.js");
+const loginCounter = require("../module/loginCounter.js");
 
 
 // 로그인
@@ -35,6 +35,7 @@ router.post("/log-in",async(req,res,next)=>{
             result.message = "로그인 성공"
             result.token = await verify.publishToken(row[0])
             result.isadmin = row[0].isadmin
+            loginCounter.countLogin(parseInt(await row[0].usernum))
         } else{
             result.message = "해당하는 회원정보가 없습니다."
         }
@@ -56,17 +57,22 @@ router.get("/log-out",async(req,res,next)=>{
     const result = {
         "success" : false,
         "message" : "",
-        "token" : null
+        "auth" : false
     }
     try{
-        // 토큰 정지 코드
+        if(!req.decoded) throw new Error('authorization Fail');
+        result.auth = true
         result.success = true
+        //res.clearCookie("token")
+        //토큰 블랙리스트(blacklist) 처리
     }catch(err){
         console.log("POST /account/log-out", err.message)
         result.message = err.message
     } finally{
-        req.resData = result
-        next() // 세션이 아닌 다른 접근...?
+        res.send(result)
+
+        req.resData = result //for logging
+        next()
     }    
 })
 
@@ -101,8 +107,10 @@ router.get("/id-exist",async(req,res,next)=>{ //아이디 중복체크
         console.log("GET /account/id-exist/",err.message)
         result.message = err.message
     }finally{
-        if(client) client.end() 
-        req.resData = result
+        if(client) client.end()
+        res.send(result)
+
+        req.resData = result //for logging
         next()
     }
     
@@ -143,9 +151,10 @@ router.post("/",async(req,res,next)=>{
         console.log("POST /account",err.message)
         result.message = err.message
     }finally{
-        if(client) client.end() 
-        
-        req.resData = result
+        if(client) client.end()
+        res.send(result)
+
+        req.resData = result //for logging
         next()
     }
 })
@@ -185,9 +194,10 @@ router.get("/find-id",async(req,res,next)=>{
         console.log("GET /account/find-id",err.message)
         result.message = err.message
     }finally {
-        if(client)client.end()
+        if(client) client.end()
+        res.send(result)
 
-        req.resData = result
+        req.resData = result //for logging
         next()
     }
         
@@ -198,7 +208,8 @@ router.get("/certification",async(req,res,next)=>{
     const {id,name,mail} = req.query; // 받아옴
     const result = {
         "success" : false,
-        "message" : ""
+        "message" : "",
+        "token":null
     }
     let client = null
     try{
@@ -218,15 +229,10 @@ router.get("/certification",async(req,res,next)=>{
     
             const row = data.rows
             if(row.length != 0) {
-                req.customData.userNum = await row[0].usernum
-                req.customData.userId = await id
                 result.success  = true
                 result.message = "귀하의 아이디를 찾았습니다."
+                result.token = await verify.publishToken(row[0]) // 임시토큰 발행
             } else{
-                if (req.customData.userNum || req.customData.userId) { //세션정보가 존재하는 경우
-                    delete req.customData.userNum
-                    delete req.customData.userId
-                }
                 result.message = "존재하지 않는 정보입니다."
             }
         }
@@ -234,8 +240,10 @@ router.get("/certification",async(req,res,next)=>{
         console.log("GET /account/certification",err.message)
         result.message = err.message
     } finally{
-        if(client)client.end()
-        req.resData = result
+        if(client) client.end()
+        res.send(result)
+
+        req.resData = result //for logging
         next()
     }
    
@@ -245,20 +253,24 @@ router.put("/modify-pw",async(req,res,next)=>{
     const {newpw1,newpw2} = req.body; // 받아옴
     const result = {
         "success" : false,
-        "message" : ""
+        "message" : "",
+        "auth" : false
     }
     let client = null
     try{
+        if(!req.decoded) throw new Error('authorization Fail');
+        result.auth = true
+
         const pwCheck = new inputCheck(newpw1)
         if(pwCheck.isMinSize(4).isMaxSize(31).isSameWith(newpw2).isEmpty().result != true) result.message = pwCheck.errMessage
         else {
             client = new Client(db.pgConnect)
             client.connect()
-            const usernum = await req.customData.userNum
             const sql = "UPDATE account SET pw = $1 WHERE usernum = $2;"
-            const values = [newpw1,usernum]
+            const values = [newpw1,await req.decoded.userNum]
             const data = await client.query(sql,values)
-
+            //res.clearCookie("token")
+            //token 블랙리스트(blacklist)처리
             result.success  = true
             result.message = "비밀번호 변경 완료"
         }
@@ -266,9 +278,10 @@ router.put("/modify-pw",async(req,res,next)=>{
         console.log("PUT /account/modify-pw",err.message)
         result.message = err.message
     }finally{
-        if(client)client.end()
+        if(client) client.end()
+        res.send(result)
 
-        req.resData = result
+        req.resData = result //for logging
         next()
     }
     
@@ -278,22 +291,24 @@ router.delete("/",async(req,res,next)=>{
     const {pw} = req.body;
     const result = {
         "success" : false,
-        "message" : ""
+        "message" : "",
+        "auth" : false
     }
     let client = null
     try{
-        const pwCheck = new inputCheck(pw)
+        if(!req.decoded) throw new Error('authorization Fail');
+        result.auth = true
 
+        const pwCheck = new inputCheck(pw)
         if(pwCheck.isMinSize(4).isMaxSize(31).isEmpty().result != true) result.message = pwCheck.errMessage
         else {
             client = new Client(db.pgConnect)
             client.connect()
-            const usernum = await req.customData.userNum
             const sql = "DELETE FROM account WHERE usernum = $1 AND pw = $2;"
-            const values = [usernum,pw]
+            const values = [req.decoded.userNum,pw]
             const data = await client.query(sql,values)
-
-            req.customData.destroy(function(err){})
+            // res.clearCookie("token")
+            // token 블랙리스트(blacklist처리) 
 
             result.success  = true
             result.message = "계정 삭제 완료"
@@ -303,8 +318,9 @@ router.delete("/",async(req,res,next)=>{
         result.message = err.message
     } finally{
         if(client) client.end()
+        res.send(result)
 
-        req.resData = result
+        req.resData = result //for logging
         next()
     }
 })
